Add CLIArgumentsParser.parse method
The parsing logic was extracted from
CLIArgumentsParser.__init__ into the "parse"
method. This way you can add new arguments without
having to modify the initializer.

The "parse"
method returns an object of type
argparse.Namespace so that you can access each CLI
argument by its name (as in args.ports, for
example), instead of accessing a dictionary by its
keys. This difference is visible in the
initialization of instance attributes in
PortScanner.__init__

The variadic arguments ""*args, **kwargs" in the
signature of CLIArgumentsParser.parse allow you
to test the class without the necessity of a TTY.