Convert the "target" CLI argument into positional
The argparse library considers an argument of the
type "--target" to be optional, which in our case
may break the execution at runtime. Since we
always need a target anyway we can just provide
the "target" string as an argument to
"add_argument" and argparse will enforce the
presence of a target anyway.

Now we can execute the script as in
```port-scanner.py scanme.nmap.org -p 80, 443```