Decouple PortScanner.__init__ arguments
The way it was a PortScanner instance could only
be initialized if we had an appropriate object
of type CLIArgumentsParser, since the
PortScanner.__init__ method took an argument of
type argparse.Namespace. What if we needed
arguments to come from a file or maybe even a
socket connection? That's why the PortScanner
initializer needs to specify its own arguments
that are then decomposed by the caller before
instantiation.

This is a good practice. Our
PortScanner class doesn't need to have any
knowledge whatsoever of where the arguments come
from and in which format. Just supply all
arguments as strings and it will just work.